refactor(web): hoist shared entity table helpers from profile routes
The LinkRow / trendYearColumns / networkColumns / networkRows block was duplicated verbatim (~55 lines) between authority.tsx and company.tsx. Extract into lib/entity-tables and import from both; pure refactor, identical rendered output.
