+++ apps/web/app/routes/authority.tsx
@@ -1,6 +1,5 @@
 import { Link } from 'react-router';
-import type { NetworkData, TrendYear } from '@sigma/api-contract';
-import { count, money, moneyBare, pct, periodRange, plural, signedPct } from '@sigma/shared';
+import { count, money, moneyBare, pct, periodRange, plural } from '@sigma/shared';
 import {
   authorityIdFromSlug,
   getAuthority,
@@ -13,7 +12,7 @@ import { Breadcrumbs } from '../components/Breadcrumbs';
 import { PageHeader } from '../components/PageHeader';
 import { FactsList } from '../components/FactsList';
 import { StackedBar } from '../components/StackedBar';
-import { DataTable, type Column } from '../components/DataTable';
+import { DataTable } from '../components/DataTable';
 import { TrendChart } from '../components/TrendChart';
 import { NetworkGraph } from '../components/NetworkGraph';
 import { ContractMiniTable } from '../components/ContractMiniTable';
@@ -21,6 +20,7 @@ import { SingleOfferPortion } from '../components/SingleOfferPortion';
 import { ShareBar, Chip, Section } from '../components/ui';
 import { publicCache } from '../lib/cache';
 import { coverageRange, getCoverageMeta } from '../lib/coverage';
+import { networkColumns, networkRows, trendYearColumns } from '../lib/entity-tables';
 import { withDbRetry } from '../lib/retry';
 import { seoMeta } from '../lib/meta';
 
@@ -37,64 +37,6 @@ export function meta({ data, params, matches }: Route.MetaArgs) {
 
 export function headers() {
   return { 'Cache-Control': publicCache(3600) };
-}
-
-interface LinkRow {
-  from: string;
-  to: string;
-  valueEur: number;
-  contracts: number;
-}
-
-const trendYearColumns: Column<TrendYear>[] = [
-  {
-    key: 'year',
-    header: 'Година',
-    isTitle: true,
-    cell: (r) => (
-      <>
-        {r.year}
-        {r.partial && <span className="muted"> (частично)</span>}
-      </>
-    ),
-  },
-  { key: 'value', header: 'Стойност', align: 'money', cell: (r) => money(r.valueEur) },
-  { key: 'contracts', header: 'Договори', align: 'num', cell: (r) => count(r.contracts) },
-  {
-    key: 'yoy',
-    header: 'Спрямо предходната',
-    align: 'num',
-    cell: (r) => (r.yoyPct == null ? '' : signedPct(r.yoyPct)),
-  },
-];
-
-const networkColumns: Column<LinkRow>[] = [
-  { key: 'from', header: 'От', isTitle: true, cell: (r) => r.from },
-  { key: 'to', header: 'Към', cell: (r) => r.to },
-  { key: 'value', header: 'Стойност', align: 'money', cell: (r) => money(r.valueEur) },
-  {
-    key: 'contracts',
-    header: 'Договори',
-    align: 'num',
-    secondary: true,
-    cell: (r) => count(r.contracts),
-  },
-];
-
-function networkRows(data: NetworkData): LinkRow[] {
-  const nodeById = new Map(data.nodes.map((n) => [n.id, n] as const));
-  return data.edges.map((e) => {
-    const a = nodeById.get(e.from);
-    const b = nodeById.get(e.to);
-    const authority = a?.kind === 'authority' ? a : b;
-    const company = a?.kind === 'authority' ? b : a;
-    return {
-      from: authority?.label ?? e.from,
-      to: company?.label ?? e.to,
-      valueEur: e.valueEur,
-      contracts: e.contracts,
-    };
-  });
 }
 
 export async function loader({ params, context }: Route.LoaderArgs) {
